Permit P10

+++ Backend/backend.Core/DTOs/LicenseApplicationUserDetailsDto.cs
@@ -12,7 +12,8 @@ namespace backend.Core.DTOs
 
         //public long RegNumber { get; set; }
 
-        public long RegId { get; set; }
+        //public long RegId { get; set; }
+         public int RegId { get; set; }
         public string? Mobile { get; set; }
         public string ApplicantName { get; set; } = string.Empty;
         //public string CompanyName { get; set; } = string.Empty;
